Refactor to split Image into Image and Frame classes
Instead of an array of Image instances, now there is a single Image instance
which contains an array of Frame instances.  This allows for things like a
single global palette for all sub-images, metadata/flags on images, and
tilesets where all sub-images must be the same dimensions (as well as those
where each sub-image can have different dimensions).